asanzo suggested fixes
