+++ locales/es-ar/creator.json
@@ -20,7 +20,10 @@
 	},
 	"statement": {
 		"title": "Enunciado",
+		"button": "¡Escribí tu Enunciado!",
 		"description": "Descripción",
-		"clue": "Pista"
+		"descriptionHint": "Escribí un enunciado para el ejercicio. Que sea conciso y diga claramente el objetivo. Podés usar **negritas**, _cursivas_ y emojis :wink: con la sintaxis de Markdown",
+		"clue": "Pista",
+		"includeClue": "Incluir una pista"
 	}
 	}
