+++ src/components/creator/Selection.tsx
@@ -6,7 +6,7 @@ import { useTranslation } from "react-i18next"
 import styles from "./selection.module.css"
 import { HomeCard } from "../home/HomeCard"
 import { LocalStorage } from "../../localStorage"
-import { Scene, SceneType, SerializedChallenge } from "../serializedChallenge"
+import { SceneType, defaultChallenge } from "../serializedChallenge"
 
 type CharacterCardProps = {
 	name: SceneType
@@ -40,28 +40,6 @@ const characters: CharacterCardProps[] = [
 	},
 ]
 
-const defaultScene = (type: SceneType): Scene => {
-	return {
-		type: type,
-		maps: [[['A', '-', '-'],
-		['-', '-', '-'],
-		['-', '-', '-']]]
-	}
-}
-
-export const defaultChallenge = (type: SceneType): SerializedChallenge => {
-	return {
-		fileVersion: 1,
-		title: "",
-		statement: {
-			description: ""
-		},
-		scene: defaultScene(type),
-		toolbox: {
-			blocks: []
-		}
-	}
-}
 
 const CharacterCard = (props: CharacterCardProps) => {
 	const { t } = useTranslation("creator")
